fix(db): widen chart_entries play counts to BigInteger
Last.fm global play counts exceed PostgreSQL's 32-bit integer range
(e.g. 3.7e9). chart_entries.playcount/listeners were declared Integer,
which SQLite tolerates but PostgreSQL rejects on insert. Widen both to
BigInteger.

The SQLite->Postgres migration script now drops + recreates the schema
instead of create_all + truncate, so a re-run always rebuilds tables
with the current column types.

+++ app/models/db.py
@@ -573,8 +573,8 @@ class ChartEntry(Base):
     track_title = Column(String(512), nullable=True)  # null for artist charts
     artist_name = Column(String(512), nullable=False)
     artist_mbid = Column(String(64), nullable=True)
-    playcount = Column(Integer, nullable=True)
-    listeners = Column(Integer, nullable=True)
+    playcount = Column(BigInteger, nullable=True)
+    listeners = Column(BigInteger, nullable=True)
 
     # Image URL from Last.fm (best available size)
     image_url = Column(String(1024), nullable=True)
